requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# PDF Processing
pymupdf==1.23.8

# Document Processing
python-docx==1.1.0
pywin32==306  # For reading legacy .doc files on Windows

# LLM Integration
openai==1.10.0
google-generativeai==0.8.0

# Utilities
python-dotenv==1.0.0
